noir_lsp::notifications
Function process_workspace_for_single_file_change

pub(crate) fn process_workspace_for_single_file_change(
    state: &mut LspState,
    workspace: &Workspace,
    file_uri: Url,
    file_source: &str,
) -> Result<(), Error>

Type-checks a single file that changed by using existing cached data for the workspace/package, such as the cached NodeInterner, CrateGraph and DefMaps.

This greatly improves the responsiveness of the LSP server when editing files. However, the cost is a slight decrease in autocompletion accuracy. For example, if a struct is removed from the code, it will still existing in the cached data and it will still be offered as autocompletion. Or for example if the compiler refuses to re-process a trait because it’s already defined, new methods defined on that trait won’t be available for autocompletion. However, this is solved when the file is saved as the entire package is re-processed then. We think this is an acceptable trade-off to improve responsiveness, and it could eventually be further improved.