noir_lsp
Module requests

Modules

code_action ğŸ”’

code_lens_request ğŸ”’

completion ğŸ”’

document_symbol ğŸ”’

expand ğŸ”’

folding_range ğŸ”’

goto_declaration ğŸ”’

goto_definition ğŸ”’

hover ğŸ”’

inlay_hint ğŸ”’

references ğŸ”’

rename ğŸ”’

semantic_tokens ğŸ”’

Handles LSP semantic tokens requests.

signature_help ğŸ”’

std_source_code ğŸ”’

test_run ğŸ”’

tests ğŸ”’

workspace_symbol ğŸ”’

Structs

ChainingHintsOptions ğŸ”’

ClosingBraceHintsOptions ğŸ”’

InlayHintsOptions ğŸ”’

LspInitializationOptions ğŸ”’

LSP client will send initialization request after the server has started. InitializeParams.initialization_options will contain the options sent from the client.

ParameterHintsOptions ğŸ”’

ProcessRequestCallbackArgs ğŸ”’

TraitReexport ğŸ”’

Represents a trait reexported from a given module with a name.

TypeHintsOptions ğŸ”’

Functions

character_to_line_offset ğŸ”’

default_chaining_hints ğŸ”’

default_chaining_hints_enabled ğŸ”’

default_closing_brace_hints ğŸ”’

default_closing_brace_hints_enabled ğŸ”’

default_closing_brace_min_lines ğŸ”’

default_enable_code_actions ğŸ”’

default_enable_code_lens ğŸ”’

default_enable_completions ğŸ”’

default_enable_inlay_hints ğŸ”’

default_enable_lightweight_mode ğŸ”’

default_enable_parsing_cache ğŸ”’

default_enable_semantic_tokens ğŸ”’

default_enable_signature_help ğŸ”’

default_inlay_hints ğŸ”’

default_parameter_hints ğŸ”’

default_parameter_hints_enabled ğŸ”’

default_type_hints ğŸ”’

default_type_hints_enabled ğŸ”’

find_all_references ğŸ”’

find_all_references_in_workspace ğŸ”’

get_reference_name ğŸ”’

on_formatting ğŸ”’

on_formatting_inner ğŸ”’

on_initialize ğŸ”’

on_shutdown ğŸ”’

position_to_byte_index ğŸ”’

position_to_location ğŸ”’

process_request ğŸ”’

process_request_impl ğŸ”’

process_request_no_type_check ğŸ”’

process_request_no_workspace_cache ğŸ”’

read_format_config ğŸ”’

semantic_token_types ğŸ”’

semantic_token_types_map ğŸ”’

to_lsp_location ğŸ”’