noir_lsp::requests

Module completion

Modules

auto_import 🔒

builtins 🔒

completion_items 🔒

kinds 🔒

params 🔒

If the cursor is at the end of a function parameter name, suggest parameter names (and their types) that exists in the same module, impl or trait.

sort_text 🔒

tests 🔒

Structs

NodeFinder 🔒

Functions

get_array_element_type 🔒

get_field_type 🔒

get_type_type_id 🔒

module_def_id_from_reference_id 🔒

name_matches 🔒

Returns true if name matches a prefix written in code. prefix must already be in snake case. This method splits both name and prefix by underscore, then checks that every part of name starts with a part of prefix, in order.

on_completion_request 🔒