noir_lsp::requests::completion::params

Function find_function_and_parameter_name_at_byte_index

fn find_function_and_parameter_name_at_byte_index<'a>(
    functions: impl Iterator<Item = &'a NoirFunction>,
    byte_index: usize,
) -> Option<(&'a NoirFunction, &'a str)>

Tries to find a function parameter inside functions that is being autocompleted. Returns that function together with the parameter name, if found.