noir_lsp::requests::completion

Module sort_text

Functions

auto_import_sort_text 🔒

Sort text for auto-import items. We want these to show up after local definitions.

crate_or_module_sort_text 🔒

We want crates and modules to show up after other things (for example local variables, functions or types)

default_sort_text 🔒

This is the default sort text.

local_variable_sort_text 🔒

Local variables should be suggested before anything else.

new_sort_text 🔒

Sort text for “new” methods: we want these to show up before anything else, if we are completing at something like Foo::

operator_sort_text 🔒

We want to show operator methods last.

self_mismatch_sort_text 🔒

When completing something like Foo::, we want to show methods that take self after the other ones.

underscore_sort_text 🔒

If a name begins with underscore it’s likely something that’s meant to be private (but visibility doesn’t exist everywhere yet, so for now we assume that)