noir_lsp::requests::completion
Struct NodeFinder

struct NodeFinder<'a> {

    files: &'a FileMap,
    file: FileId,
    source: &'a str,
    lines: Vec<&'a str>,
    byte_index: usize,
    byte: Option<u8>,
    module_id: ModuleId,
    def_maps: &'a BTreeMap<CrateId, CrateDefMap>,
    dependencies: &'a Vec<Dependency>,
    interner: &'a NodeInterner,
    completion_items: Vec<CompletionItem>,
    local_variables: HashMap<String, Span>,
    type_parameters: HashSet<String>,
    suggested_module_def_ids: HashSet<ModuleDefId>,
    nesting: usize,
    auto_import_line: usize,
    use_segment_positions: UseSegmentPositions,
    self_type: Option<Type>,
    in_comptime: bool,
    func_id: Option<FuncId>,

}

Fields
files: &'a FileMapfile: FileIdsource: &'a strlines: Vec<&'a str>byte_index: usizebyte: Option<u8>module_id: ModuleId

The module ID in scope. This might change as we traverse the AST if we are analyzing something inside an inline module declaration.

def_maps: &'a BTreeMap<CrateId, CrateDefMap>dependencies: &'a Vec<Dependency>interner: &'a NodeInternercompletion_items: Vec<CompletionItem>

Completion items we find along the way.

local_variables: HashMap<String, Span>

Local variables in the current scope, mapped to their locations. As we traverse the AST, we collect local variables.

type_parameters: HashSet<String>

Type parameters in the current scope. These are collected when entering a struct, a function, etc., and cleared afterwards.

suggested_module_def_ids: HashSet<ModuleDefId>

ModuleDefIds we already suggested, so we don’t offer these for auto-import.

nesting: usize

How many nested mod we are in deep

auto_import_line: usize

The line where an auto_import must be inserted

use_segment_positions: UseSegmentPositionsself_type: Option<Type>in_comptime: boolfunc_id: Option<FuncId>

The function we are in, if any

Implementations

impl NodeFinder<'_>

pub(super) fn complete_auto_imports( &mut self, prefix: &str, requested_items: RequestedItems, function_completion_kind: FunctionCompletionKind, )

fn get_ancestor_module_reexport( &self, module_def_id: ModuleDefId, visibility: ItemVisibility, ) -> Option<Reexport>

impl NodeFinder<'_>

pub(super) fn builtin_functions_completion( &mut self, prefix: &str, function_completion_kind: FunctionCompletionKind, )

pub(super) fn builtin_values_completion(&mut self, prefix: &str)

pub(super) fn builtin_types_completion(&mut self, prefix: &str)

pub(super) fn suggest_builtin_attributes( &mut self, prefix: &str, target: AttributeTarget, )

fn suggest_allow(&mut self, name: &'static str, prefix: &str)

impl NodeFinder<'_>

pub(super) fn module_def_id_completion_items( &self, module_def_id: ModuleDefId, name: String, function_completion_kind: FunctionCompletionKind, function_kind: FunctionKind<'_>, requested_items: RequestedItems, ) -> Vec<CompletionItem>

pub(super) fn crate_completion_item( &self, name: impl Into<String>, id: ModuleId, ) -> CompletionItem

pub(super) fn module_completion_item( &self, name: impl Into<String>, id: ModuleId, ) -> CompletionItem

fn struct_completion_item( &self, name: String, type_id: TypeId, ) -> CompletionItem

fn enum_completion_item(&self, name: String, type_id: TypeId) -> CompletionItem

pub(super) fn struct_field_completion_item( &self, field: &str, typ: &Type, struct_id: TypeId, field_index: usize, self_type: bool, ) -> CompletionItem

fn type_alias_completion_item( &self, name: String, id: TypeAliasId, ) -> CompletionItem

fn trait_associated_type_completion_item( &self, name: String, id: TraitAssociatedTypeId, ) -> CompletionItem

fn trait_completion_item( &self, name: String, trait_id: TraitId, ) -> CompletionItem

fn global_completion_item( &self, name: String, global_id: GlobalId, ) -> CompletionItem

pub(super) fn enum_variant_completion_item( &self, name: String, type_id: TypeId, variant_index: usize, ) -> CompletionItem

pub(super) fn function_completion_items( &self, name: &String, func_id: FuncId, function_completion_kind: FunctionCompletionKind, function_kind: FunctionKind<'_>, attribute_first_type: Option<&Type>, trait_info: Option<(TraitId, Option<&TraitReexport>)>, self_prefix: bool, ) -> Vec<CompletionItem>

pub(super) fn function_completion_item( &self, name: &String, func_id: FuncId, func_meta: &FuncMeta, func_self_type: Option<&Type>, function_completion_kind: FunctionCompletionKind, function_kind: FunctionKind<'_>, attribute_first_type: Option<&Type>, trait_info: Option<(TraitId, Option<&TraitReexport>)>, self_prefix: bool, is_macro_call: bool, ) -> CompletionItem

fn auto_import_trait_if_trait_method( &self, func_id: FuncId, trait_info: Option<(TraitId, Option<&TraitReexport>)>, completion_item: &mut CompletionItem, ) -> Option<()>

fn compute_function_insert_text( &self, func_meta: &FuncMeta, name: &str, function_kind: FunctionKind<'_>, skip_first_argument: bool, ) -> String

pub(super) fn completion_item_with_doc_comments( &self, id: ReferenceId, completion_item: CompletionItem, ) -> CompletionItem

fn hir_pattern_to_argument(&self, pattern: &HirPattern, text: &mut String)

fn hir_pattern_is_self_type(&self, pattern: &HirPattern) -> bool

impl<'a> NodeFinder<'a>

fn new( files: &'a FileMap, file: FileId, source: &'a str, byte_index: usize, byte: Option<u8>, krate: CrateId, def_maps: &'a BTreeMap<CrateId, CrateDefMap>, dependencies: &'a Vec<Dependency>, interner: &'a NodeInterner, ) -> Self

fn find(&mut self, parsed_module: &ParsedModule) -> Option<CompletionResponse>

fn complete_constructor_field_name( &mut self, constructor_expression: &ConstructorExpression, )

fn find_in_path(&mut self, path: &Path, requested_items: RequestedItems)

fn find_in_path_impl( &mut self, path: &Path, requested_items: RequestedItems, in_the_middle: bool, )

fn local_variables_completion(&mut self, prefix: &str)

fn type_parameters_completion(&mut self, prefix: &str)

fn find_in_use_tree(&mut self, use_tree: &UseTree, prefixes: &mut Vec<Path>)

fn find_in_use_tree_path( &mut self, prefixes: &Vec<Path>, ident: &Ident, alias: &Option<Ident>, )

fn collect_local_variables(&mut self, pattern: &Pattern)

fn collect_type_parameters_in_generics(&mut self, generics: &UnresolvedGenerics)

fn collect_type_parameters_in_generic(&mut self, generic: &UnresolvedGeneric)

fn complete_type_fields_and_methods( &mut self, typ: &Type, prefix: &str, function_completion_kind: FunctionCompletionKind, self_prefix: bool, )

fn complete_type_methods( &mut self, typ: &Type, prefix: &str, function_kind: FunctionKind<'_>, function_completion_kind: FunctionCompletionKind, self_prefix: bool, )

fn complete_trait_constraints_methods( &mut self, typ: &Type, prefix: &str, function_kind: FunctionKind<'_>, function_completion_kind: FunctionCompletionKind, )

fn complete_trait_methods( &mut self, trait_: &Trait, prefix: &str, function_kind: FunctionKind<'_>, function_completion_kind: FunctionCompletionKind, )

fn complete_enum_variants_without_parameters( &mut self, data_type: &DataType, prefix: &str, )

fn complete_struct_fields( &mut self, struct_type: &DataType, generics: &[Type], prefix: &str, self_prefix: bool, )

fn complete_tuple_fields(&mut self, types: &[Type], self_prefix: bool)

fn complete_in_module( &mut self, module_id: ModuleId, prefix: &str, path_kind: PathKind, at_root: bool, function_completion_kind: FunctionCompletionKind, requested_items: RequestedItems, )

fn resolve_module(&self, segments: Vec<Ident>) -> Option<ModuleId>

fn resolve_path(&self, segments: Vec<Ident>) -> Option<ModuleDefId>

fn suggest_no_arguments_attributes(&mut self, prefix: &str, attributes: &[&str])

fn suggest_one_argument_attributes(&mut self, prefix: &str, attributes: &[&str])

fn suggest_trait_impl_function( &mut self, noir_trait_impl: &NoirTraitImpl, noir_function: &NoirFunction, )

fn try_set_self_type(&mut self, pattern: &Pattern)

fn get_lvalue_type(&self, lvalue: &LValue) -> Option<Type>

fn complete_module_declaration( &mut self, module: &ModuleDeclaration, ) -> Option<()>

Try to suggest the name of a module to declare based on which files exist in the filesystem, excluding modules that are already declared.

fn module_def_id_is_visible( &self, module_def_id: ModuleDefId, visibility: ItemVisibility, defining_module: Option<ModuleId>, ) -> bool

fn includes_span(&self, span: Span) -> bool