noir_lsp::requests::hover
Module from_reference

Structs

TypeLinksGatherer ğŸ”’

Functions

append_doc_comments ğŸ”’

append_value_to_string ğŸ”’

format_alias ğŸ”’

format_enum ğŸ”’

format_enum_variant ğŸ”’

format_function ğŸ”’

format_generic_names ğŸ”’

format_generics ğŸ”’

format_generics_impl ğŸ”’

format_global ğŸ”’

format_link ğŸ”’

format_local ğŸ”’

format_module ğŸ”’

format_parent_module ğŸ”’

format_parent_module_from_module_id ğŸ”’

format_pattern ğŸ”’

format_reference ğŸ”’

format_struct ğŸ”’

format_struct_member ğŸ”’

format_trait ğŸ”’

format_trait_associated_type ğŸ”’

format_type ğŸ”’

get_trait_impl_func_id ğŸ”’

go_to_type_links ğŸ”’

hover_from_reference ğŸ”’

link_target_location ğŸ”’

Returns the Location where a link target exists. The Location might not exist. For example, primitive types have no definition location.

pattern_is_self ğŸ”’

process_doc_comments_links ğŸ”’

Replaces markdown links inside doc comments that point to Noir items to point at code locations where those items are defined.

value_to_string ğŸ”’