noir_lsp

Module types

Modules

notification 🔒

request 🔒

Structs

CodeLens 🔒

A code lens represents a command that should be shown along with source text, like the number of references, a way to run tests, etc.

CodeLensOptions 🔒

Code Lens options.

CodeLensParams 🔒

Command 🔒

Represents a reference to a command. Provides a title which will be used to represent a command in the UI. Commands are identified by a string identifier. The recommended way to handle commands is to implement their execution on the server side if the client and server provides the corresponding capabilities. Alternatively the tool extension code could handle the command. The protocol currently doesn’t specify a set of well-known commands.

Diagnostic 🔒

Represents a diagnostic, such as a compiler error or warning. Diagnostic objects are only valid in the scope of a resource.

DiagnosticSeverity 🔒

The protocol currently supports the following diagnostic severities:

DidChangeConfigurationParams 🔒

DidChangeTextDocumentParams 🔒

DidCloseTextDocumentParams 🔒

DidOpenTextDocumentParams 🔒

DidSaveTextDocumentParams 🔒

InitializeParams 🔒

InitializeResult 🔒

InitializedParams 🔒

NargoCapability 🔒

NargoExpandParams 🔒

NargoPackageTests 🔒

NargoStdSourceCodeParams 🔒

NargoTest 🔒

NargoTestId 🔒

NargoTestRunParams 🔒

NargoTestRunResult 🔒

NargoTestsOptions 🔒

NargoTestsParams 🔒

Position 🔒

Position in a text document expressed as zero-based line and character offset. A position is between two characters like an ‘insert’ cursor in a editor.

PublishDiagnosticsParams 🔒

Range 🔒

A range in a text document expressed as (zero-based) start and end positions. A range is comparable to a selection in an editor. Therefore the end position is exclusive.

ServerCapabilities 🔒

ServerInfo 🔒

Url 🔒

A parsed URL record.

Enums

TextDocumentSyncCapability 🔒

Type Aliases

CodeLensResult 🔒

GotoDeclarationResult 🔒

GotoDefinitionResult 🔒

NargoExpandResult 🔒

NargoStdSourceCodeResult 🔒

NargoTestsResult 🔒