noir_lsp::types

Module notification

Structs

NargoUpdateTests 🔒

Enums

DidChangeConfiguration 🔒

A notification sent from the client to the server to signal the change of configuration settings.

DidChangeTextDocument 🔒

The document change notification is sent from the client to the server to signal changes to a text document. In 2.0 the shape of the params has changed to include proper version numbers and language ids.

DidCloseTextDocument 🔒

The document close notification is sent from the client to the server when the document got closed in the client. The document’s truth now exists where the document’s uri points to (e.g. if the document’s uri is a file uri the truth now exists on disk).

DidOpenTextDocument 🔒

The document open notification is sent from the client to the server to signal newly opened text documents. The document’s truth is now managed by the client and the server must not try to read the document’s truth using the document’s uri.

DidSaveTextDocument 🔒

The document save notification is sent from the client to the server when the document was saved in the client.

Exit 🔒

A notification to ask the server to exit its process. The server should exit with success code 0 if the shutdown request has been received before; otherwise with error code 1.

Initialized 🔒

The initialized notification is sent from the client to the server after the client received the result of the initialize request but before the client is sending any other request or notification to the server. The server can use the initialized notification for example to dynamically register capabilities.