noir_lsp::types
Module request

Structs

Initialize 🔒

NargoExpand 🔒

NargoStdSourceCode 🔒

NargoTestRun 🔒

NargoTests 🔒

Enums

CodeLens 🔒

The code lens request is sent from the client to the server to compute code lenses for a given text document.

Formatting 🔒

The document formatting request is sent from the server to the client to format a whole document.

GotoDeclaration 🔒

GotoDefinition 🔒

The goto definition request is sent from the client to the server to resolve the definition location of a symbol at a given text document position.

GotoTypeDefinition 🔒

The goto type definition request is sent from the client to the server to resolve the type definition location of a symbol at a given text document position.

Shutdown 🔒

The shutdown request is sent from the client to the server. It asks the server to shut down, but to not exit (otherwise the response might not be delivered correctly to the client). There is a separate exit notification that asks the server to exit.